AI/econe/requirements.txt
django
pandas
joblib
scikit-learn
djangorestframework
drf-yasg
setuptools
